play-extraction-poc/requirements.txt
anthropic>=0.34.0
Pillow>=10.0.0
pypdf>=4.0.0
pytest>=7.0.0
pyyaml>=6.0
cairosvg>=2.7.0
